Codes/requirements.txt
streamlit
requests
openai
scikit-learn
google-generativeai
